requirements.txt
streamlit
matplotlib
numpy
pillow
easyocr
requests
pandas
